+++ tests/antithesis/workloads/v1/bank/anytime_disable_faults.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-touch /faults-disabled
-echo "Precept faults disabled"