+++ tests/antithesis/workloads/v1/bank/finally_validate.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-/test_client --remote s3-compatible bank-validate